+++ packages/core/src/components/Table/Foot/types.ts
@@ -1,0 +1,5 @@
+import { TableProps } from '../types';
+
+export interface TableFootProps {
+    tableSize: TableProps['size'];
+}