+++ packages/core/src/types.ts
@@ -1,4 +1,5 @@
+import { DateRangeSelectionEnum } from './components/DateRangePicker/types';
 import { TableColumnConfig, TableProps } from './components/Table/types';
 import { ToastVariants } from './components/Toast/types';
 
-export { TableColumnConfig, TableProps, ToastVariants };
+export { TableColumnConfig, TableProps, ToastVariants, DateRangeSelectionEnum };