+++ packages/layout/src/components/Header/NavItem/NavItem.styled.tsx
@@ -18,6 +18,7 @@ export const NavItem = styled.button<NavItemProps>`
     align-items: center;
     text-align: left;
     grid-auto-columns: max-content;
+    cursor: pointer;
 
     ${({ theme, isActive }) => media(breakpoints(theme.breakpoints).up('L'))`
         background: ${theme.header.backgroundColor};
@@ -32,7 +33,7 @@ export const NavItem = styled.button<NavItemProps>`
             content: '';
             position: absolute;
             width: calc(100% - 1.6rem * 2);
-            top: calc((${theme.header.navItem.height.desktop} - ${theme.header.height}) / 2);
+            top: calc((${theme.header.navItem.height.desktop} - ${theme.header.height.desktop}) / 2);
             left: 1.6rem;
             background: ${theme.header.navItem.activeIndicatorColor};
             height: 4px;